Djangorestframework/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
filelock==3.13.1
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
pipenv==2023.12.1
platformdirs==4.2.0
PyAudio==0.2.14
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2024.1
pywin32==306
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.1.0
six==1.16.0
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.0
Werkzeug==3.0.1
wikipedia==1.4.0
